+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## [0.5.75] - 2020-01-07
+### Added
+- The `date format`, `datetime format`, and `time format`
+  interview-wide defaults.
+### Fixed
+- Line breaking issue with `terms` and `auto terms`.
+
 ## [0.5.74] - 2020-01-06
 ### Fixed
 - Error when checkbox values are not strings.